# docs/openapi/open-api.yaml
openapi: 3.0.3
info:
  title: AGI 开放 API
  version: v1
  description: |
    供数字档案馆等第三方业务系统调用的 AGI 平台开放接口。

    - 认证：`X-AGI-App-Code` + `X-AGI-Api-Key`
    - 上下文（可选）：`X-AGI-User-Id`、`X-AGI-Unit-Id`、`X-AGI-Department-Ids`、`X-AGI-Role-Ids`
    - 响应格式：统一 `ApiResponse` 信封
servers:
  - url: /
    description: 当前环境
...
security:
  - AppCode: []
    ApiKey: []
paths:
...
  /api/open/bots/{id}:
    get:
      tags: [智能体]
      summary: 获取智能体详情
      operationId: getBot
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BotResponse'
  /api/open/bots/{id}/run:
    post:
      tags: [智能体]
      summary: 运行智能体
      operationId: runBot
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OpenBotRunRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BotRunResponse'
  /api/open/bots/{id}/chat:
    post:
      tags: [智能体]
      summary: 多轮对话（同步）
      operationId: chatBot
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OpenBotChatRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BotChatResponse'
  /api/open/bots/{id}/capabilities:
    get:
      tags: [智能体]
      summary: 列出智能体多工作流路由能力
      operationId: listBotCapabilities
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BotCapabilityPageResponse'
  /api/open/bots/{id}/sessions:
    get:
      tags: [智能体]
      summary: 列出用户会话
      description: 需携带 X-AGI-User-Id 或请求体 userId
      operationId: listBotSessions
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BotSessionPageResponse'
    post:
      tags: [智能体]
      summary: 创建会话
      operationId: createBotSession
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OpenBotCreateSessionRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BotSessionResponse'
  /api/open/bots/{id}/sessions/{sessionId}/messages:
    get:
      tags: [智能体]
      summary: 列出会话消息
      operationId: listBotMessages
      parameters:
        - $ref: '#/components/parameters/ResourceId'
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BotMessagePageResponse'
  /api/open/bots/{id}/sessions/{sessionId}/messages/stream:
    post:
      tags: [智能体]
      summary: 多轮对话（SSE 流式）
      description: |
        SSE 事件类型：message.delta、message.completed、citation.added、tool.completed、
        confirm.required、job.progress、job.completed、error、done
      operationId: streamBotMessage
      parameters:
        - $ref: '#/components/parameters/ResourceId'
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OpenBotStreamRequest'
      responses:
        '200':
          description: SSE 流
          content:
            text/event-stream:
              schema:
                type: string
...
  /api/open/identity/resolve:
    post:
      tags: [身份]
      summary: 解析调用身份上下文
      operationId: resolveIdentity
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResolveIdentityRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdentityResponse'
components:
  securitySchemes:
    AppCode:
      type: apiKey
      in: header
      name: X-AGI-App-Code
    ApiKey:
      type: apiKey
      in: header
      name: X-AGI-Api-Key
    UserId:
      type: apiKey
      in: header
      name: X-AGI-User-Id
    UnitId:
      type: apiKey
      in: header
      name: X-AGI-Unit-Id
    DepartmentIds:
      type: apiKey
      in: header
      name: X-AGI-Department-Ids
    RoleIds:
      type: apiKey
      in: header
      name: X-AGI-Role-Ids
  parameters:
    ResourceId:
      name: id
      in: path
      required: true
      schema:
        type: string
  schemas:
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
        data: {}
        error:
          $ref: '#/components/schemas/ErrorResponse'
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        requestId:
          type: string
        details:
          type: object
          additionalProperties: true
    PageResponse:
      type: object
      properties:
        items:
          type: array
          items: {}
        total:
          type: integer
          format: int64
    OpenBotView:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        description: { type: string }
        avatar: { type: string }
        workflowId: { type: string }
        modelProviderId: { type: string }
        knowledgeBaseIds:
          type: array
          items: { type: string }
        openingMessage: { type: string }
        capabilityHint: { type: string }
        suggestedQuestions:
          type: array
          items: { type: string }
        status: { type: string }
    OpenBotCapabilityView:
      type: object
      properties:
        id: { type: string }
        capabilityType: { type: string }
        capabilityId: { type: string }
        capabilityCode: { type: string }
        routingKeywords: { type: string }
        primaryCapability: { type: boolean }
        enabled: { type: boolean }
    OpenBotCreateSessionRequest:
      type: object
      properties:
        title: { type: string }
        userId: { type: string }
    OpenBotStreamRequest:
      allOf:
        - $ref: '#/components/schemas/OpenBotRunRequest'
        - type: object
          properties:
            message: { type: string }
    BotSession:
      type: object
      properties:
        id: { type: string }
        botId: { type: string }
        title: { type: string }
        messageCount: { type: integer }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        pinned: { type: boolean }
    BotMessage:
      type: object
      properties:
        id: { type: string }
        sessionId: { type: string }
        botId: { type: string }
        role: { type: string, enum: [USER, ASSISTANT] }
        content: { type: string }
        messageType: { type: string }
        metadata:
          type: object
          additionalProperties: true
        createdAt: { type: string, format: date-time }
    OpenBotChatResponseData:
      type: object
      properties:
        session:
          $ref: '#/components/schemas/BotSession'
        messages:
          type: array
          items:
            $ref: '#/components/schemas/BotMessage'
        reply:
          $ref: '#/components/schemas/BotMessage'
        execution:
          $ref: '#/components/schemas/WorkflowExecutionResponse'
    OpenBotRunRequest:
      type: object
      properties:
        message: { type: string }
        userId: { type: string }
        unitId: { type: string }
        departmentIds:
          type: array
          items: { type: string }
        roleIds:
          type: array
          items: { type: string }
        input:
          type: object
          additionalProperties: true
    OpenBotChatRequest:
      allOf:
        - $ref: '#/components/schemas/OpenBotRunRequest'
        - type: object
          properties:
            sessionId: { type: string }
...
    WorkflowExecutionResponse:
      type: object
      properties:
        id: { type: string }
        workflowId: { type: string }
        workflowVersionId: { type: string }
        status: { type: string }
        input:
          type: object
          additionalProperties: true
        output:
          type: object
          additionalProperties: true
        errorMessage: { type: string }
        startedAt: { type: string, format: date-time }
        finishedAt: { type: string, format: date-time }
...
    ResolveIdentityRequest:
      type: object
      properties:
        unitId: { type: string }
        departmentIds:
          type: array
          items: { type: string }
        roleIds:
          type: array
          items: { type: string }
        userId: { type: string }
    RuntimeIdentityContext:
      type: object
      properties:
        tenantId: { type: string }
        appId: { type: string }
        userId: { type: string }
        unitIds:
          type: array
          items: { type: string }
        activeUnitId: { type: string }
        departmentIds:
          type: array
          items: { type: string }
        roleIds:
          type: array
          items: { type: string }
        authType: { type: string }
        source: { type: string }
...
    BotResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/OpenBotView'
    BotRunResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
    BotChatResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/OpenBotChatResponseData'
    BotCapabilityPageResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              allOf:
                - $ref: '#/components/schemas/PageResponse'
                - type: object
                  properties:
                    items:
                      type: array
                      items:
                        $ref: '#/components/schemas/OpenBotCapabilityView'
    BotSessionPageResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              allOf:
                - $ref: '#/components/schemas/PageResponse'
                - type: object
                  properties:
                    items:
                      type: array
                      items:
                        $ref: '#/components/schemas/BotSession'
    BotSessionResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/BotSession'
    BotMessagePageResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              allOf:
                - $ref: '#/components/schemas/PageResponse'
                - type: object
                  properties:
                    items:
                      type: array
                      items:
                        $ref: '#/components/schemas/BotMessage'
...
    IdentityResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/RuntimeIdentityContext'